A partial re-write of the coordinator/agents.
The coordinator now assigns jobs to agents (in memory) and then agents come "pick up" their jobs that they've been assigned. Previously, the coordinator had no representation of the full set of agents and would just wait for an agent to request a job and choose a job for the agent at that time. This new implementation allows the coordinator to be a little more strategic in deciding which jobs should go to which agent (see _create_workers_for_job, this addresses a previous TODO), but the real impetus is to prepare for future changes that will allow the coordinator to create new agents (e.g. in AWS/Azure) to run jobs. In order to do this, the coordinator needs to know what the universe of available agents is and it needs to be able to "refrain" from sending jobs to existing agents if we are waiting for new agents for that job to be spun up in AWS/Azure. Related minor changes:
- Renaming job_worker to agent to reduce confusion.
- Agents register themselves with the coordinator
- get_next_job is now get_next_jobs so that the coordinator can give an agent more than one job at a time
- Moved management of resources from the agent to the coordinator
- Fleshed out coordinator's representation of agents/grid workers, and moved data classes and logic into resource_allocation.py
- Get rid of the ASSIGNED state, no longer necessary now that we have a richer representation of the state of agents/grid workers.

A related major change is to address some TODOs relating to grid job failures.
- The first is that previously, if a grid_worker failed to start up (e.g. we can't get the docker container image) the coordinator would get a job update with state=RUN_REQUEST_FAILED that would be ignored. We've addressed this by just failing the next unassigned task in that grid job (see update_grid_job_state).
- Second, previously, if a grid_worker failed while running a task, the coordinator would get a job update with state=NON_ZERO_RETURN_CODE, but that would be ignored. We've addressed this by giving the grid workers ids, and having the coordinator keep track of which grid worker is working on which task, which allows the coordinator to fail the appropriate task. This also sets the stage for implementing grid worker timeouts in the future.
- Third, previously, if there a job's required resources prevented it from running on any agent, the job would not fail and no error messages would be thrown. We've addressed this by having the coordinator just fail every task in the grid job/fail the simple job in this scenario.